Macro mutually_exclusive_features::exactly_one_of

macro_rules! exactly_one_of {
    ($($F:literal),+ $(,)?) => { ... };
    (@comma_sep: $last:literal $(,)?) => { ... };
    (@comma_sep: $head:literal, $($rest:literal),+ $(,)?) => { ... };
}

This is exactly like none_or_one_of except that it requires exactly one of the features to be selected.

To use it, simply call it with the features you want to be mutually exclusive:

use mutually_exclusive_features::exactly_one_of;
exactly_one_of!("feature1", "feature2", "feature3");

Which will generate the following code:

#[cfg(all(feature="feature1", feature="feature2"))]
compile_error!("The `feature1` and `feature2` features are mutually exclusive and cannot be enabled at the same time!");

#[cfg(all(feature="feature1", feature="feature3"))]
compile_error!("The `feature1` and `feature3` features are mutually exclusive and cannot be enabled at the same time!");

#[cfg(all(feature="feature2", feature="feature3"))]
compile_error!("The `feature2` and `feature3` features are mutually exclusive and cannot be enabled at the same time!");

#[cfg(not(any(feature="feature1", feature="feature2", feature="feature3")))]
compile_error!("You must enable exactly one of `feature1`, `feature2`, `feature3` features!");